Rename `LinearAlgebra` object to `linearalgebra.Types`
