+++ src/main/scala/scotty/simulator/math/linearalgebra/Types.scala
@@ -1,9 +1,9 @@
-package scotty.simulator.math
+package scotty.simulator.math.linearalgebra
 
 import org.apache.commons.math3.complex.Complex
 import org.apache.commons.math3.linear.{Array2DRowFieldMatrix, ArrayFieldVector}
 
-object LinearAlgebra {
+object Types {
   type ApacheMatrix = Array2DRowFieldMatrix[Complex]
   type ApacheVector = ArrayFieldVector[Complex]
 }
